backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
aiomysql==0.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.0
beautifulsoup4==4.12.3
boto3==1.34.155
celery==5.3.4
certifi==2022.6.15
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.1.1
configparser==5.3.0
cryptography==38.0.1
daphne==4.0.0
Deprecated==1.2.13
Django==4.1
django-admin-list-filter-dropdown==1.0.3
django-celery-beat==2.5.0
django-cors-headers==3.14.0
django-filter==22.1
django-oauth-toolkit==2.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-ses==4.3.0
ez-setup==0.9
frozenlist==1.3.3
gunicorn
idna==3.3
importlib-metadata==4.12.0
jwcrypto==1.4.2
kombu==5.4.2
Markdown==3.4.1
moviepy==1.0.3
multidict==6.0.4
mysql-connector==2.2.9
mysqlclient==2.1.1
oauthlib==3.2.1
openai==0.28.1
paytmchecksum==1.7.0
psycopg2==2.9.5
pycparser==2.21
pycryptodome==3.17
PyJWT==2.4.0
pytz==2022.2.1
pillow==10.4.0
requests==2.28.1
razorpay==1.4.2
six==1.16.0
sqlparse==0.4.2
textblob==0.18.0
tqdm==4.64.1
urllib3==1.26.12
uvicorn[standard]
vaderSentiment==3.3.2
wrapt==1.14.1
yarl==1.8.2
zipp==3.8.1
